fixing features broken by merge
